Add an updateUser method to change linked accounts (use ref instead of reactive)
