+++ apps/web/src/composables/auth.ts
@@ -1,10 +1,6 @@
-import { ProviderString } from '@/types/ProviderString'
 import { LoginCredentials } from '@casimir/types'
-import { reactive } from 'vue'
 
 export default function useAuth() {
-    const usersAccounts = reactive<any>({})
-
     /**
      * Logs a user in with an address, message and signed message
      * 
@@ -36,51 +32,8 @@ export default function useAuth() {
         }
     }
 
-    function addAccount(provider: ProviderString, address: string) {
-        address = address.toLowerCase()
-        const localStorage = window.localStorage
-        const accounts = JSON.parse(localStorage.getItem('accounts') as string) || {}
-
-        for (const existingProvider in accounts) {
-            if (accounts[existingProvider].includes(address) && existingProvider !== provider) {
-                accounts[existingProvider] = accounts[existingProvider].filter((existingAddress: string) => existingAddress !== address)
-            }
-        }
-
-        if (!accounts[provider] && address) {
-            accounts[provider] = [address]
-        } else if (address) {
-            if (!accounts[provider].includes(address)) {
-                accounts[provider].push(address)
-            }
-        }
-
-        for (const provider in accounts) {
-            usersAccounts[provider] = accounts[provider]
-        }
-
-        localStorage.setItem('accounts', JSON.stringify(accounts))
-    }
-
-    function removeAccount(provider: ProviderString, address: string) {
-        address = address.toLowerCase()
-        const localStorage = window.localStorage
-        const accounts = JSON.parse(localStorage.getItem('accounts') as string) || {}
-        
-        if (accounts[provider] && address) {
-            accounts[provider] = accounts[provider].filter((account: string) => account !== address)
-        }
-
-        for (const provider in accounts) {
-            usersAccounts[provider] = accounts[provider]
-        }
-        localStorage.setItem('accounts', JSON.stringify(accounts))
-    }
-
     return {
         login,
-        addAccount,
-        removeAccount,
         usersAccounts
     }
 }
