+++ apps/web/src/interfaces/User.ts
@@ -1,0 +1,6 @@
+import { ProviderString } from '@/types/ProviderString'
+
+export interface User {
+    id: string;
+    accounts: Record<ProviderString, string[]>;
+}
